Do not use upstream_node_ips check
