+++ tests/schains/checks_test.py
@@ -10,12 +10,16 @@ import mock
 import docker
 import pytest
 
+from skale.schain_config.generator import get_schain_nodes_with_schains
+
+
 from core.schains.checks import SChainChecks, CheckRes
 from core.schains.config.file_manager import UpstreamConfigFilename
 from core.schains.config.directory import (
     get_schain_check_filepath,
     schain_config_dir
 )
+from core.schains.config.schain_node import generate_schain_nodes
 from core.schains.skaled_exit_codes import SkaledExitCodes
 from core.schains.runner import get_container_info, get_image_name, run_ima_container
 # from core.schains.cleaner import remove_ima_container
@@ -25,7 +29,13 @@ from tools.helper import read_json
 
 from web.models.schain import upsert_schain_record, SChainRecord
 
-from tests.utils import CONFIG_STREAM, get_schain_contracts_data, response_mock, request_mock
+from tests.utils import (
+    CONFIG_STREAM,
+    generate_schain_config,
+    get_schain_contracts_data,
+    response_mock,
+    request_mock
+)
 
 
 NOT_EXISTS_SCHAIN_NAME = 'qwerty123'
@@ -125,7 +135,7 @@ def test_dkg_check(schain_checks, sample_false_checks):
     assert not sample_false_checks.dkg.status
 
 
-def test_upstream_config_check(schain_checks):
+def test_upstream_config_check(skale, schain_checks):
     assert not schain_checks.upstream_config
     ts = int(time.time())
     name, rotation_id = schain_checks.name, schain_checks.rotation_id
@@ -135,8 +145,18 @@ def test_upstream_config_check(schain_checks):
         f'schain_{name}_{rotation_id}_{ts}.json'
     )
 
+    schain_nodes_with_schains = get_schain_nodes_with_schains(skale, name)
+    nodes = generate_schain_nodes(
+        schain_nodes_with_schains=schain_nodes_with_schains,
+        schain_name=name,
+        rotation_id=rotation_id
+    )
+
+    config = generate_schain_config(name)
+    config['skaleConfig']['sChain']['nodes'] = nodes
+
     with open(upstream_path, 'w') as upstream_file:
-        json.dump({'config': 'upstream'}, upstream_file)
+        json.dump(config, upstream_file)
 
     assert schain_checks.upstream_config
